backend/requirements/core.txt
Django==3.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.0.0